+++ examples/main/src/Components/SocialPollsServiceMock/poll-mocks-data.ts
@@ -1,0 +1,46 @@
+const poll = {
+  id: '24035f02-b39e-47e0-bbdc-242242f22c69',
+  title: 'Is RiCos the GREATEST?',
+  mediaId: 'https://static.wixstatic.com/media/436483e6ed9e41fe91b9f286d2ea4efb.jpg',
+  createdBy: 'd0d683f9-81b1-4ec2-84ee-7f49c5245148',
+  count: 0,
+  anonymousCount: 0,
+  ownVotes: [],
+  settings: {
+    multipleVotes: false,
+    voteRole: 'SITE_MEMBERS',
+    resultsVisibility: 'ALWAYS',
+    votersDisplay: true,
+    votesDisplay: true,
+  },
+  options: [
+    {
+      id: '91756cfb-d47a-4dd1-9eda-df847981016b',
+      title: 'Yes',
+      mediaId: 'https://static.wixstatic.com/media/3a27828e768344158489df4f2c03cfb2.jpg',
+      count: 0,
+      anonymousCount: 0,
+      rating: 0,
+      latestVoters: [],
+    },
+    {
+      id: '64fb4cf2-b379-4be6-97fc-79cbc0355ccd',
+      title: 'HELL YES!!',
+      mediaId: 'https://static.wixstatic.com/media/fb62d9f8e7824b2e90b13b971e700b92.jpg',
+      count: 0,
+      anonymousCount: 0,
+      rating: 0,
+      latestVoters: [],
+    },
+  ],
+  creatorFlag: true,
+};
+
+const voters = [
+  {
+    memberId: 'd0d683f9-81b1-4ec2-84ee-7f49c5245148',
+    voted: '2021-10-06T08:27:25.967Z',
+  },
+];
+
+export { poll, voters };