+++ packages/plugin-social-polls/web/src/toolbar/createToolbar.ts
@@ -108,6 +108,7 @@ export const createToolbar: CreatePluginToolbar = ({ isMobile, settings, t }) =>
               t,
               activeTab: TABS.EDIT,
               mobile: true,
+              settings,
             },
 
             {
@@ -119,6 +120,7 @@ export const createToolbar: CreatePluginToolbar = ({ isMobile, settings, t }) =>
               t,
               activeTab: TABS.DESIGN,
               mobile: true,
+              settings,
             },
           ]
         : [
@@ -133,6 +135,7 @@ export const createToolbar: CreatePluginToolbar = ({ isMobile, settings, t }) =>
               tooltipTextKey: 'Poll_PollSettings_Common_Header',
               activeTab: TABS.LAYOUT,
               mobile: false,
+              settings,
             },
 
             { keyName: 'separator', mobile: false, type: BUTTONS.SEPARATOR },
@@ -148,6 +151,7 @@ export const createToolbar: CreatePluginToolbar = ({ isMobile, settings, t }) =>
               tooltipTextKey: 'Poll_PollSettings_Common_Header',
               activeTab: TABS.DESIGN,
               mobile: false,
+              settings,
             },
 
             { keyName: 'separator', mobile: false, type: BUTTONS.SEPARATOR },
@@ -163,6 +167,7 @@ export const createToolbar: CreatePluginToolbar = ({ isMobile, settings, t }) =>
               tooltipTextKey: 'Poll_FormatToolbar_Settings_Tooltip',
               activeTab: TABS.SETTINGS,
               mobile: false,
+              settings,
             },
           ]),